# docs/swagger/swagger.json
{
  "basePath": "/",
  "definitions": {
    "V1BearerAuthError": {
      "properties": {
        "error": {
          "properties": {
            "message": {
              "example": "unauthenticated",
              "type": "string"
            }
          },
          "required": [
            "message"
          ],
          "type": "object"
        }
      },
      "required": [
        "error"
      ],
      "type": "object"
    },
...
    "V1StandardError": {
      "properties": {
        "error": {
          "properties": {
            "code": {
              "example": "invalid_json",
              "type": "string"
            },
            "message": {
              "type": "string"
            }
          },
          "required": [
            "code",
            "message"
          ],
          "type": "object"
        }
      },
      "required": [
        "error"
      ],
      "type": "object"
    },
...
  },
  "host": "localhost:8080",
  "info": {
    "contact": {
      "name": "AVF Engineering",
      "url": "https://github.com/avf/avf-vending-system"
    },
    "description": "HTTP API for the AVF vending platform (cmd/api). All `/v1/*` routes require `Authorization: Bearer <JWT>` unless noted (health, optional metrics, Swagger when enabled). Bearer middleware failures use minimal JSON `{\"error\":{\"message\":\"...\"}}` without `error.code`. Most `/v1` handlers use `{\"error\":{\"code\":\"...\",\"message\":\"...\"}}` (see writeAPIError). **501** list stubs return `not_implemented` plus `capability` and `implemented:false`. **503** `capability_not_configured` is used when optional wiring (MQTT dispatch, commerce persistence, outbox defaults) is missing. Request tracing: responses echo `X-Request-ID` and `X-Correlation-ID` when middleware is enabled.",
    "license": {
      "name": "License terms apply per your deployment agreement; see repository NOTICE if present."
    },
    "termsOfService": "https://github.com/avf/avf-vending-system/tree/main/avf-vending-api",
    "title": "AVF Vending HTTP API",
    "version": "1.0"
  },
  "paths": {
...
    "/v1/admin/organizations/{orgId}/artifacts": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "Organization UUID",
            "in": "path",
            "name": "orgId",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "json"
        ],
        "responses": {
          "200": {
            "description": "List artifacts",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "401": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "403": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List artifacts",
        "tags": [
          "Artifacts"
        ]
      },
      "post": {
        "consumes": [
          "json"
        ],
        "description": "Routes are **not mounted** unless API_ARTIFACTS_ENABLED=true and object store is configured. Requires org_admin for orgId or platform_admin. Uses sensitive-write rate limiting when HTTP_RATE_LIMIT_SENSITIVE_WRITES_ENABLED=true \u2192 **429** `rate_limited`.",
        "parameters": [
          {
            "description": "Organization UUID",
            "in": "path",
            "name": "orgId",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "json"
        ],
        "responses": {
          "201": {
            "description": "Reserve artifact id",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "401": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "403": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "429": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Reserve artifact id",
        "tags": [
          "Artifacts"
        ]
      }
    },
    "/v1/admin/organizations/{orgId}/artifacts/{artifactId}": {
      "delete": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "Organization UUID",
            "in": "path",
            "name": "orgId",
            "required": true,
            "type": "string"
          },
          {
            "description": "Artifact UUID",
            "in": "path",
            "name": "artifactId",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "json"
        ],
        "responses": {
          "200": {
            "description": "Delete artifact",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "401": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "403": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "404": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "429": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Delete artifact",
        "tags": [
          "Artifacts"
        ]
      },
      "get": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "Organization UUID",
            "in": "path",
            "name": "orgId",
            "required": true,
            "type": "string"
          },
          {
            "description": "Artifact UUID",
            "in": "path",
            "name": "artifactId",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "json"
        ],
        "responses": {
          "200": {
            "description": "Get artifact metadata",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "401": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "403": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "404": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get artifact metadata",
        "tags": [
          "Artifacts"
        ]
      }
    },
    "/v1/admin/organizations/{orgId}/artifacts/{artifactId}/content": {
      "put": {
        "consumes": [
          "octet-stream"
        ],
        "description": "Request body is raw bytes. Required headers: **Content-Length**; optional **Content-Type**, **X-Artifact-SHA256**, **X-Artifact-Filename**. Sensitive-write rate limit may return **429**.",
        "parameters": [
          {
            "description": "Organization UUID",
            "in": "path",
            "name": "orgId",
            "required": true,
            "type": "string"
          },
          {
            "description": "Artifact UUID",
            "in": "path",
            "name": "artifactId",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "json"
        ],
        "responses": {
          "200": {
            "description": "Upload artifact bytes",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "401": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "403": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "404": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "429": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Upload artifact bytes",
        "tags": [
          "Artifacts"
        ]
      }
    },
    "/v1/admin/organizations/{orgId}/artifacts/{artifactId}/download": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "Organization UUID",
            "in": "path",
            "name": "orgId",
            "required": true,
            "type": "string"
          },
          {
            "description": "Artifact UUID",
            "in": "path",
            "name": "artifactId",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "json"
        ],
        "responses": {
          "200": {
            "description": "Presigned download URL",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "401": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "403": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "404": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Presigned download URL",
        "tags": [
          "Artifacts"
        ]
      }
    },
...
  },
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "description": "Send: Bearer <JWT>. JWT mode is selected by HTTP_AUTH_MODE (hs256 default, rs256_pem, rs256_jwks). 401/403 responses from this layer use JSON {\"error\":{\"message\":\"...\"}} without error.code.",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0",
...
}